API actions and functions now injected to client class
Also, they now perform the action or function when called instead
of creating the query data or string. Actions can be found under the
"actions" parameter, and functions under the "functions" parameter.
The "action" parameter can still be used to store the action
or function if the provided ones do not have the one you need.